Small polishing for GitHub repository search.
